refactor(1013): extract PromptClientUtils facade (SC-001, position 35)
Extract PromptClientUtils (213 lines, 31 refs) from MistHelper.py into
src/input/prompt_client_utils.py under the SC-001 facade pattern used
throughout Cat B initiative 1013.

Cat B position 35. Live-global reads (apisession, InputUtils, PromptUtils,
ConfigUtils) resolve via lazy `mh = importlib.import_module("MistHelper")`
inside each helper. Direct imports cover stdlib + installed packages
(mistapi, prettytable). Callers continue to reach the class through the
MistHelper.PromptClientUtils re-export alias.

- Local gate: black --check + ruff + py_compile CLEAN
- MistHelper.py --test baseline: 66/0/131 preserved
