+++ pyproject.toml
@@ -248,6 +248,7 @@ omit = [
     "src/export/sites_by_ap_model_exporter.py",
     "src/firmware/firmware_manager.py",
     "src/gateway/gateway_ha_exporter.py",
+    "src/input/prompt_client_utils.py",
     "src/inventory/org_device_inventory_summary_facade.py",
     "src/reports/sfp_transceiver_data_processor.py",
     "src/reports/wired_client_manufacturer_report_generator.py",
