+++ src/input/__init__.py
@@ -1,0 +1,1 @@
+"""Interactive prompt utilities package."""
